requirements.txt
Django==5.0.*
djangorestframework==3.15.*
psycopg[binary]==3.2.*
gunicorn==22.0.*
pytest==8.2.*
pytest-django==4.8.*